Implement basic file-backed caching
Caching now writes a set of Elisp code to a customizable file location.
Implement basic sorting.
The cache is sorted to avoid re-sorting on every UI re-render.

The cache implementation is now storing the plist instead of the rows.
This allows nicer modification of the cache and decouples cache from UI.